Add Metadata

+++ Tests/CreateAPITests/Expected/petstore-custom-imports/Sources/Paths.swift
@@ -76,11 +76,11 @@ private struct QueryParameterEncoder {
         ISO8601DateFormatter().string(from: value)
     }
 
-    static func encode(_ value: Int) -> String? {
+    static func encode(_ value: Double) -> String? {
         String(value)
     }
 
-    static func encode(_ value: Double) -> String? {
+    static func encode(_ value: Int) -> String? {
         String(value)
     }
